app/requirements.txt
python-dotenv
discord.py
aiohttp
emoji